requirements.txt
##
## basic: env, network, type
python-dotenv~=1.0.1
httpx~=0.28.1
pydantic~=2.10.5
asyncio~=3.4.3

## AI
qianfan

##
## langchain
langchain~=0.3.15
langchain-community~=0.3.15
langchain-openai~=0.3.1
langchain-core~=0.3.31
langchain-text-splitters~=0.3.5

##
## doc
docx2txt~=0.8
pypdf~=5.1.0
extract_msg~=0.52.0
PyPDF2~=3.0.1

##
## vector db
faiss-cpu~=1.10.0

## gui
streamlit~=1.42.0
pandas~=2.2.3
numpy~=2.2.2
st-pages

## Rest
flask~=3.1.0
flask-restx~=1.3.0
flask-httpauth~=4.8.0
Werkzeug~=3.1.3

##
## test
pytest~=8.3.4
requests~=2.32.3